MCM

Documentation for MCM.

MCM.convert_rate_mcm

MCM.evaluate_rate

MCM.extract_mechanism

MCM.mechanism_to_catalyst

MCM.convert_rate_mcm — Method

convert_rate_mcm(rate_string::String)

Convert an MCM rate expression to a valid Julia expression we can evaluate after defining necessary constants.

MCM.evaluate_rate — Method

evaluate_rate(rate::String)

Evaluate the MCM reaction rate using simulation parameters.

MCM.extract_mechanism — Method

extract_mechanism(filename::String)

Reach the mechaism equation file, extracting the reaction strings and rate coefficients for later ingestion into Catalyst.jl

MCM.mechanism_to_catalyst — Method

mechanism_to_catalyst(rxns, species_list, out_file, rn_name)

Take the output of extractmechanism() and generate a julia file called outfile defining a reaction network called rn_name using the Catalyst.jl DSL.